requirements.txt
fastapi
uvicorn[standard]
httpx
gtfs-realtime-bindings
python-dotenv
tzdata
